requirements.txt
PyExecJS2==1.6.1
Requests==2.32.3
tabulate==0.9.0
